Enabling some image drawing tests, moving image drawing to a seperate file and updating some image file names
