+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGContext.RedBox.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:591b4932e3d1585f38b0cae499d9ba89c500a66eab6d26c80b2c447fa9e9493d
+size 11186
