+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGImageDrawing.DrawAContextIntoAnImage.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:33d915050f05e75cf606d02438d99fb9a1741e92dffbd503c7e5f617411fe87b
+size 10310
