+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGImageDrawing.DrawAnImage.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:991fad944afe1b47b2b6d7495ade0cd5dccd584c37a2372d86c01e42cbe7e4a5
+size 577337
