+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGImageDrawing.TiledImageCustom.png
