+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGImageDrawing.TiledImageCustomScaledUp.png
