+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGImageDrawing.TiledImageFlowerScaledDown.png
