+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGImageDrawing.TiledImageMultiColorFlower.png
