+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGImageDrawing.TiledImageMultiColorFlowerScaled.png
