+++ tests/unittests/CoreGraphics.drawing/data/tiledImageFlower.png
