+++ tests/unittests/CoreGraphics.drawing/data/tiledImageMultiColorFlower.png
